refactor to Python package
