+++ Dockerfile
@@ -8,10 +8,9 @@ COPY Pipfile.lock /
 WORKDIR /
 RUN pipenv install --system --deploy --ignore-pipfile
 
-COPY *.py /
-COPY templates /templates
+COPY kube_resource_report /kube_resource_report
 
 ARG VERSION=dev
-RUN sed -i "s/__version__ = .*/__version__ = '${VERSION}'/" /report.py
+RUN sed -i "s/__version__ = .*/__version__ = '${VERSION}'/" /kube_resource_report/__init__.py
 
-ENTRYPOINT ["/report.py"]
+ENTRYPOINT ["python3", "-m", "kube_resource_report"]
