+++ kube_resource_report/__init__.py
@@ -1,0 +1,1 @@
+__version__ = '0.1-3-g00a7914-dirty'
