+++ kube_resource_report/templates/applications.html
