+++ kube_resource_report/templates/base.html
