+++ kube_resource_report/templates/cluster.html
