+++ kube_resource_report/templates/clusters.html
