+++ kube_resource_report/templates/favicon.png
