+++ kube_resource_report/templates/index.html
