+++ kube_resource_report/templates/ingresses.html
