+++ kube_resource_report/templates/kube-resource-report.css
