+++ kube_resource_report/templates/pods.html
