+++ kube_resource_report/templates/sortable-theme-minimal.css
