+++ kube_resource_report/templates/team.html
