+++ kube_resource_report/templates/teams.html
