+++ sample-report/generate-report.sh
@@ -11,9 +11,9 @@ kubectl apply -f ../deploy/
 
 VERSION=$(git describe --tags --always --dirty)
 echo "Updating version to $VERSION.."
-sed -i "s/__version__ = .*/__version__ = '${VERSION}'/" ../report.py
+sed -i "s/__version__ = .*/__version__ = '${VERSION}'/" ../kube_resource_report/__init__.py
 
 sleep 10
 
-(cd ../ && OAUTH2_ACCESS_TOKENS=read-only=mytok pipenv run ./report.py sample-report/output/ --application-registry $(minikube service application-registry --url) \
+(cd ../ && OAUTH2_ACCESS_TOKENS=read-only=mytok pipenv run python3 -m kube_resource_report sample-report/output/ --application-registry $(minikube service application-registry --url) \
     --additional-cost-per-cluster=30 --kubeconfig-contexts=minikube)
